# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline Labs API
  version: 2.0.0
  description: |
    Pipeline Labs is an SDK-first platform for orchestrating LLM and ML training workflows
    on infrastructure owned by the customer. The API supports Supabase bearer tokens for the web app
    and server-issued API keys for SDK and automation use cases.
servers:
  - url: https://pipelinelabs-ashen.vercel.app/api
    description: Production
  - url: http://localhost:3001/api
    description: Local development
...
paths:
...
  /planner/analyze:
    post:
      tags: [Planner]
      summary: Analyze dataset and generate training plan
      operationId: analyzeDatasetForTraining
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                datasetId:
                  type: string
                  format: uuid
                userObjective:
                  type: string
              required: [datasetId]
      responses:
        '200':
          description: Analysis and generated plan
          content:
            application/json:
              schema:
                type: object
                properties:
                  analysis:
                    $ref: '#/components/schemas/DatasetAnalysis'
                  plan:
                    $ref: '#/components/schemas/TrainingPlan'
                  planId:
                    type: string
                    format: uuid
        '404':
          $ref: '#/components/responses/NotFound'
  /planner:
    get:
      tags: [Planner]
      summary: List training plans
      operationId: listTrainingPlans
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Training plans for the current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingPlanRecord'
  /planner/{planId}:
    get:
      tags: [Planner]
      summary: Get training plan
      operationId: getTrainingPlan
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Training plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingPlanRecord'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Planner]
      summary: Update training plan
      operationId: updateTrainingPlan
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan:
                  $ref: '#/components/schemas/TrainingPlan'
                status:
                  type: string
                  enum: [draft, approved, rejected, executing, completed, failed]
      responses:
        '200':
          description: Updated training plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingPlanRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token used by the web app.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Server-issued API key for SDK and automation workflows.
...
  schemas:
...
    DatasetAnalysis:
      type: object
      properties:
        data_type:
          type: string
          enum: [tabular, text, image, audio, mixed]
        row_count:
          type: integer
        column_count:
          type: integer
        columns:
          type: array
          items:
            $ref: '#/components/schemas/DatasetColumn'
        label_column:
          type: string
          nullable: true
        task_type:
          type: string
        suggested_models:
          type: array
          items:
            $ref: '#/components/schemas/SuggestedModel'
        preprocessing_steps:
          type: array
          items:
            type: string
        data_quality_score:
          type: number
    DatasetColumn:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [numerical, categorical, text, datetime, id]
        unique_count:
          type: integer
        null_percent:
          type: number
        sample_values:
          type: array
          items:
            type: string
    SuggestedModel:
      type: object
      properties:
        name:
          type: string
        framework:
          type: string
        gpu_required:
          type: string
        estimated_time_minutes:
          type: number
        estimated_cost_usd:
          type: number
        confidence:
          type: number
    TrainingPlan:
      type: object
      properties:
        task:
          type: string
        model:
          type: string
        framework:
          type: string
          enum: [pytorch, tensorflow, sklearn, xgboost]
        train_test_split:
          type: string
        batch_size:
          type: integer
        optimizer:
          type: string
        learning_rate:
          type: number
        epochs:
          type: integer
        gpu_required:
          type: string
          enum: [none, T4, A10G, V100, A100]
        preprocessing:
          type: array
          items:
            type: string
        metrics:
          type: array
          items:
            type: string
        estimated_time_minutes:
          type: number
        estimated_cost_usd:
          type: number
        hyperparameter_search:
          type: object
          nullable: true
        reasoning:
          type: string
          nullable: true
    TrainingPlanRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        dataset_id:
          type: string
          format: uuid
        plan:
          $ref: '#/components/schemas/TrainingPlan'
        dataset_analysis:
          $ref: '#/components/schemas/DatasetAnalysis'
        status:
          type: string
          enum: [draft, approved, rejected, executing, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found