# server/routes/planner.ts
import { Router, Request, Response } from 'express';
import { analyzeDataset, aiEnhancedPlan } from '../services/planner';
import { getAuth } from '../auth';
import { supabaseAdmin } from '../supabase';
import Papa from 'papaparse';

export const plannerRouter = Router();

// Analyze a dataset and generate a training plan
plannerRouter.post('/analyze', async (req: Request, res: Response) => {
  try {
    const auth = getAuth(req);
    const { datasetId, userObjective } = req.body;

    if (!datasetId) {
      res.status(400).json({ error: 'datasetId is required' });
      return;
    }

    // Fetch dataset from Supabase
    const { data: dataset, error: dsError } = await supabaseAdmin
      .from('datasets')
      .select('*')
      .eq('id', datasetId)
      .eq('user_id', auth.userId)
      .maybeSingle();

    if (dsError || !dataset) {
      res.status(404).json({ error: 'Dataset not found' });
      return;
    }

    // Download dataset from storage
    const { data: storageData, error: storageError } = await supabaseAdmin.storage
      .from('datasets')
      .download(dataset.storage_path);

    if (storageError || !storageData) {
      res.status(500).json({ error: 'Failed to download dataset' });
      return;
    }

    const text = await storageData.text();
    const isJSON = dataset.storage_path?.endsWith('.json');
    let data: Record<string, unknown>[];

    if (isJSON) {
      data = JSON.parse(text);
      if (!Array.isArray(data)) data = [data];
    } else {
      // Use papaparse for robust CSV parsing
      const parsed = Papa.parse(text, {
        header: true,
        dynamicTyping: true,
        skipEmptyLines: true,
      });
      data = parsed.data as Record<string, unknown>[];
    }

    // Step A: Analyze dataset
    const analysis = analyzeDataset(data);

    // Step B: Generate training plan (AI-enhanced if possible)
    const geminiKey = process.env.GEMINI_API_KEY;
    const plan = await aiEnhancedPlan(analysis, data.slice(0, 10), userObjective, geminiKey);

    // Save to database
    const { data: planRecord, error: planError } = await supabaseAdmin
      .from('training_plans')
      .insert({
        user_id: auth.userId,
        dataset_id: datasetId,
        plan,
        dataset_analysis: analysis,
        status: 'draft',
      })
      .select()
      .single();

    if (planError) {
      console.error('Failed to save plan:', planError);
      // Still return the plan even if save fails
    }

    // Update dataset with detected info
    const { error: updateError } = await supabaseAdmin
      .from('datasets')
      .update({
        data_type: analysis.data_type,
        label_column: analysis.label_column,
        task_type: analysis.task_type,
      })
      .eq('id', datasetId);

    if (updateError) {
      console.error('Failed to update dataset:', updateError, 'datasetId:', datasetId);
      // Non-fatal error, continue with response
    }

    res.json({
      analysis,
      plan,
      planId: planRecord?.id,
    });
  } catch (err: any) {
    console.error('[Planner] Error:', err);
    res.status(500).json({ error: err.message || 'Analysis failed' });
  }
});

// Get an existing training plan
plannerRouter.get('/:planId', async (req: Request, res: Response) => {
  try {
    const auth = getAuth(req);

    const { data, error } = await supabaseAdmin
      .from('training_plans')
      .select('*')
      .eq('id', req.params.planId)
      .eq('user_id', auth.userId)
      .maybeSingle();

    if (error || !data) {
      res.status(404).json({ error: 'Plan not found' });
      return;
    }

    res.json(data);
  } catch (err: any) {
    res.status(500).json({ error: err.message });
  }
});

// Update a training plan (user modifies before approving)
plannerRouter.patch('/:planId', async (req: Request, res: Response) => {
  try {
    const auth = getAuth(req);

    const { plan, status } = req.body;

    const updates: Record<string, unknown> = {};
    if (plan) updates.plan = plan;
    if (status) updates.status = status;

    const { data, error } = await supabaseAdmin
      .from('training_plans')
      .update(updates)
      .eq('id', req.params.planId)
      .eq('user_id', auth.userId)
      .select()
      .maybeSingle();

    if (error || !data) {
      res.status(400).json({ error: error?.message || 'Plan not found' });
      return;
    }

    res.json(data);
  } catch (err: any) {
    res.status(500).json({ error: err.message });
  }
});

// List all training plans for user
plannerRouter.get('/', async (req: Request, res: Response) => {
  try {
    const auth = getAuth(req);

    const { data, error } = await supabaseAdmin
      .from('training_plans')
      .select('*, datasets(file_name, row_count, column_count)')
      .eq('user_id', auth.userId)
      .order('created_at', { ascending: false });

    if (error) {
      res.status(400).json({ error: error.message });
      return;
    }

    res.json(data);
  } catch (err: any) {
    res.status(500).json({ error: err.message });
  }
});